Create `TransactionalTrait`
